refactor(SuperChat): split into SuperChat panel, SuperChatConversations, and SuperChatInbox
Break the monolithic SuperChat (sidebar + main panel) into three composable
components sharing one folder and one import path
(@mieweb/ui/components/SuperChat):

- SuperChat        — single-conversation panel (takes one `conversation`);
                     root data-slot="superchat".
- SuperChatConversations — the conversation list (sidebar) with controlled/
                     uncontrolled selection; root data-slot="superchat-conversations".
- SuperChatInbox   — wrapper composing the two; accepts the original full
                     SuperChatProps so it is a drop-in for the old component;
                     root data-slot="superchat-inbox".

Shared presentational pieces and helpers (ParticipantAvatar, ReferenceChip,
MessageRow, Composer, sidebarItem, time/mention utils) move to an internal
parts.tsx (not exported). No package.json/tsup change — the existing single
entry + ./components/* wildcard cover the new exports.

Stories: split per component (Panel / Conversations / Inbox) with shared
sample data in storyData.tsx; add args-driven Playground stories so the
Controls panel is populated. Tests: retarget panel tests to a single
conversation and add SuperChatConversations + SuperChatInbox coverage
(21 passing). Docs: README + MAINTAINERS remapped to the three-component
model and new data-slot vocabulary.
