+++ src/components/SuperChat/SuperChatInbox.stories.tsx
@@ -1,0 +1,411 @@
+import * as React from 'react';
+import type { Meta, StoryObj } from '@storybook/react-vite';
+import { SuperChatInbox, createMarkdownRenderer } from './index';
+import {
+  createCodePlugin,
+  createMathPlugin,
+  createGenUIPlugin,
+  createMermaidPlugin,
+  createImagePlugin,
+  createNitroTablePlugin,
+} from './plugins';
+import type { SuperChatConversation } from './index';
+import {
+  conversation,
+  richConversation,
+  secondConversation,
+  registry,
+} from './storyData';
+import 'katex/dist/katex.min.css';
+
+// ============================================================================
+// Meta
+// ============================================================================
+
+const meta: Meta<typeof SuperChatInbox> = {
+  title: 'Product/Feature Modules/SuperChat/Inbox',
+  component: SuperChatInbox,
+  tags: ['autodocs'],
+  argTypes: {
+    readOnly: {
+      control: 'boolean',
+      description: 'Disable the composer.',
+      table: { category: 'Behavior' },
+    },
+    showSidebar: {
+      control: 'boolean',
+      description: 'Show the conversation list.',
+      table: { category: 'Behavior' },
+    },
+    trustedContent: {
+      control: 'boolean',
+      description:
+        'Skip sanitization — only for host-authored content.',
+      table: { category: 'Behavior' },
+    },
+    currentParticipantId: {
+      control: 'select',
+      options: ['u1', 'u2', 'a1', 'a2'],
+      description: 'The local user id (drives alignment + compose identity).',
+      table: { category: 'Identity' },
+    },
+    defaultActiveConversationId: {
+      control: 'select',
+      options: ['c1', 'c2'],
+      description: 'Uncontrolled initial active conversation id.',
+      table: { category: 'Selection' },
+    },
+    // Complex/object + callback props are wired in code, not via controls.
+    conversations: { control: false, table: { category: 'Data' } },
+    activeConversationId: { control: false, table: { category: 'Selection' } },
+    renderPlugins: { control: false, table: { category: 'Rendering' } },
+    renderTextContent: { control: false, table: { category: 'Rendering' } },
+    linkBuilder: { control: false, table: { category: 'Rendering' } },
+    className: { control: false },
+    onMessageSent: { control: false, table: { category: 'Callbacks' } },
+    onConversationOpened: { control: false, table: { category: 'Callbacks' } },
+    onConversationClosed: { control: false, table: { category: 'Callbacks' } },
+    onNewConversation: { control: false, table: { category: 'Callbacks' } },
+    onReferenceClick: { control: false, table: { category: 'Callbacks' } },
+  },
+  parameters: {
+    layout: 'fullscreen',
+    docs: {
+      description: {
+        component: [
+          '`SuperChatInbox` is the combined surface: the `SuperChatConversations` list plus the active',
+          '`SuperChat` panel. It is the drop-in equivalent of the original monolithic component — it',
+          'accepts the full `conversations` array and owns active-conversation selection (controlled via',
+          '`activeConversationId` or uncontrolled via `defaultActiveConversationId`).',
+          '',
+          'It is a native, multi-participant chat surface: any mix of multiple **AI agents** and multiple',
+          '**humans** in one conversation, generalizing `chat-component` roles into a **participant** model.',
+          '',
+          'Use the **Playground** story with the Controls panel to toggle `readOnly`, `showSidebar`,',
+          '`trustedContent`, the current participant, and the initial conversation.',
+          '',
+          '### Rich Markdown',
+          'Message text renders through a pluggable Markdown pipeline (`createMarkdownRenderer`). Opt into',
+          '`code`, `math`, `genui`, `mermaid`, `image`, and `nitro-table` plugins from',
+          '`@mieweb/ui/components/SuperChat/plugins`.',
+        ].join('\n'),
+      },
+    },
+  },
+};
+
+export default meta;
+type Story = StoryObj<typeof SuperChatInbox>;
+
+// ============================================================================
+// Stateful demo wrapper
+// ============================================================================
+// SuperChatInbox is controlled (the host owns conversation state). This wrapper
+// shows the expected host wiring: append the sent message to the active
+// conversation's thread, and simulate a reply from any @-mentioned agent.
+
+function InteractiveInbox(
+  props: Omit<React.ComponentProps<typeof SuperChatInbox>, 'conversations'> & {
+    initial: SuperChatConversation[];
+  }
+) {
+  const { initial, ...rest } = props;
+  const [conversations, setConversations] = React.useState(initial);
+
+  const appendMessage = (
+    conversationId: string,
+    message: SuperChatConversation['thread'][number]
+  ) => {
+    setConversations((prev) =>
+      prev.map((c) =>
+        c.id === conversationId
+          ? { ...c, thread: [...c.thread, message], lastActivity: message.time }
+          : c
+      )
+    );
+  };
+
+  return (
+    <SuperChatInbox
+      {...rest}
+      conversations={conversations}
+      onMessageSent={(text, meta) => {
+        const now = new Date().toISOString();
+        appendMessage(meta.conversation.id, {
+          id: `m-${Date.now()}`,
+          participantId: props.currentParticipantId ?? 'u1',
+          text,
+          time: now,
+        });
+        // Simulate each mentioned agent replying shortly after.
+        meta.conversation.participants
+          .filter((p) => p.kind === 'agent' && meta.mentions.includes(p.id))
+          .forEach((agent, i) => {
+            window.setTimeout(
+              () =>
+                appendMessage(meta.conversation.id, {
+                  id: `a-${Date.now()}-${agent.id}`,
+                  participantId: agent.id,
+                  text: `On it — responding to **${text.slice(0, 40)}**.`,
+                  time: new Date().toISOString(),
+                }),
+              500 * (i + 1)
+            );
+          });
+      }}
+    />
+  );
+}
+
+// ============================================================================
+// Stories
+// ============================================================================
+
+export const Playground: Story = {
+  args: {
+    currentParticipantId: 'u1',
+    showSidebar: true,
+    readOnly: false,
+    trustedContent: false,
+    defaultActiveConversationId: 'c1',
+  },
+  render: (args) => (
+    <div style={{ height: '90vh' }}>
+      <InteractiveInbox
+        {...args}
+        initial={[conversation, secondConversation]}
+        linkBuilder={(ref) => `#/${ref.refType}/${ref.refId}`}
+      />
+    </div>
+  ),
+};
+
+export const MarkdownCore: Story = {
+  render: () => (
+    <div style={{ height: '90vh' }}>
+      <InteractiveInbox
+        initial={[conversation, secondConversation]}
+        currentParticipantId="u1"
+        onConversationOpened={(c) => console.log('opened', c.id)}
+        onReferenceClick={(ref) => console.log('ref', ref)}
+        linkBuilder={(ref) => `#/${ref.refType}/${ref.refId}`}
+      />
+    </div>
+  ),
+};
+
+export const WithRichPlugins: Story = {
+  render: () => (
+    <div style={{ height: '90vh' }}>
+      <InteractiveInbox
+        initial={[richConversation, secondConversation]}
+        currentParticipantId="u1"
+        renderPlugins={[
+          createCodePlugin(),
+          createMathPlugin(),
+          createGenUIPlugin(registry),
+          createMermaidPlugin(),
+          createImagePlugin(),
+          createNitroTablePlugin(),
+        ]}
+        linkBuilder={(ref) => `#/${ref.refType}/${ref.refId}`}
+      />
+    </div>
+  ),
+};
+
+export const ReadOnly: Story = {
+  render: () => (
+    <div style={{ height: '90vh' }}>
+      <SuperChatInbox
+        conversations={[richConversation]}
+        currentParticipantId="u1"
+        readOnly
+        renderPlugins={[
+          createCodePlugin(),
+          createMathPlugin(),
+          createGenUIPlugin(registry),
+          createMermaidPlugin(),
+          createImagePlugin(),
+          createNitroTablePlugin(),
+        ]}
+      />
+    </div>
+  ),
+};
+
+// ============================================================================
+// Sources & Guards
+// ============================================================================
+// Documents, per visual, the exact Markdown *source* that produces it and the
+// *guard* (trust boundary) that keeps untrusted model/agent output safe. The
+// rendered column uses the real production renderer (`createMarkdownRenderer`)
+// with every plugin enabled, so source → guard → output stays in sync with the
+// code.
+
+interface FeatureDemo {
+  /** Plugin / feature name. */
+  name: string;
+  /** Raw Markdown exactly as it arrives in a message. */
+  source: string;
+  /** Where the guard lives + what it enforces. */
+  guard: string;
+}
+
+const FEATURES: FeatureDemo[] = [
+  {
+    name: 'Markdown core (GFM)',
+    source: [
+      '**Chief complaint:** chest tightness on exertion',
+      '',
+      '- Duration: 3 days',
+      '- Risk: family history of CAD',
+      '',
+      '> Recommend prioritizing an ECG.',
+      '',
+      '[ECG protocol](https://example.org/ecg)',
+    ].join('\n'),
+    guard:
+      'rehype-sanitize allow-list (createMarkdownRenderer.tsx) strips scripts/unknown tags from untrusted output; links are forced to target="_blank" rel="noopener noreferrer".',
+  },
+  {
+    name: 'code',
+    source: [
+      '```javascript',
+      "const code = '93000';",
+      'console.log(code);',
+      '```',
+    ].join('\n'),
+    guard:
+      'rehype-highlight emits .hljs-* token classes that the base schema explicitly allow-lists on code/pre/span; sanitize runs AFTER highlight so only those classes survive. Copy uses navigator.clipboard.',
+  },
+  {
+    name: 'math (KaTeX)',
+    source: [
+      '$$ risk = \\beta_0 + \\beta_1 x + \\beta_2 x^2 $$',
+      '',
+      'Inline: $x > 0.7$.',
+    ].join('\n'),
+    guard:
+      "math.tsx allow-lists KaTeX's HTML+MathML tags/attributes so its output survives sanitize; rehype-katex runs with throwOnError:false (malformed math degrades, never throws).",
+  },
+  {
+    name: 'genui',
+    source: [
+      '```genui',
+      '{ "widget": "kpi_card", "version": 1, "props": { "label": "Risk", "value": "High", "trend": "+12%" } }',
+      '```',
+    ].join('\n'),
+    guard:
+      'Widgets are host-registered, lazy, and schema-validated; the rehype transform allow-lists only the <genui-widget> tag. Unknown/invalid widgets degrade to an inert code block; mount + data fetch are gated on streaming.',
+  },
+  {
+    name: 'mermaid',
+    source: [
+      '```mermaid',
+      'graph TD',
+      '  A[Intake] --> B{Chest pain?}',
+      '  B -- Yes --> C[Order ECG]',
+      '  B -- No --> D[Routine review]',
+      '```',
+    ].join('\n'),
+    guard:
+      "mermaid.tsx loads mermaid lazily and renders with securityLevel:'strict' (labels sanitized, scripts stripped). The SVG bypasses rehype-sanitize, so strict mode IS the trust boundary; rendering is gated on streaming.",
+  },
+  {
+    name: 'image (lightbox)',
+    source:
+      '![12-lead ECG rhythm strip](https://placehold.co/640x320/png?text=ECG+rhythm+strip)',
+    guard:
+      'The image src/alt are already protocol-restricted by rehype-sanitize; image.tsx only adds the zoom affordance and portals the LightboxModal to document.body.',
+  },
+  {
+    name: 'nitro-table',
+    source: [
+      '| Code | Description | Modifier |',
+      '| --- | --- | --- |',
+      '| 93000 | ECG, complete | — |',
+      '| 93005 | ECG, tracing only | TC |',
+    ].join('\n'),
+    guard:
+      'nitroTable.tsx lazy-loads the DataVis grid only when a table appears; a GridErrorBoundary degrades to the themed HTML table if datavis is unavailable or the grid throws.',
+  },
+];
+
+const sourcesAndGuardsRenderer = createMarkdownRenderer({
+  plugins: [
+    createCodePlugin(),
+    createMathPlugin(),
+    createGenUIPlugin(registry),
+    createMermaidPlugin(),
+    createImagePlugin(),
+    createNitroTablePlugin(),
+  ],
+});
+
+function SourcesAndGuardsDemo() {
+  return (
+    <div className="space-y-6 p-6 text-neutral-900 dark:text-neutral-100">
+      <header className="space-y-1">
+        <h2 className="text-lg font-semibold">
+          SuperChat — Sources &amp; Guards
+        </h2>
+        <p className="max-w-prose text-sm text-neutral-600 dark:text-neutral-400">
+          Each visual below is produced from the raw Markdown{' '}
+          <strong>source</strong> and protected by the <strong>guard</strong>{' '}
+          noted underneath. The rendered column uses the production
+          <code> createMarkdownRenderer</code> with every plugin enabled.
+        </p>
+      </header>
+
+      {FEATURES.map((f) => (
+        <section
+          key={f.name}
+          className="rounded-xl border border-neutral-200 p-4 dark:border-neutral-700"
+        >
+          <h3 className="text-primary-700 dark:text-primary-300 mb-3 text-sm font-semibold tracking-wide">
+            {f.name}
+          </h3>
+
+          <div className="mb-1 text-xs font-medium text-neutral-500 uppercase">
+            Source
+          </div>
+          <pre className="mb-3 overflow-x-auto rounded-lg bg-neutral-900 p-3 text-xs whitespace-pre-wrap text-neutral-100 **:wrap-break-word dark:bg-neutral-950">
+            <code>{f.source}</code>
+          </pre>
+
+          <div className="mb-1 text-xs font-medium text-neutral-500 uppercase">
+            Rendered
+          </div>
+          <div className="rounded-lg border border-neutral-200 p-3 dark:border-neutral-700">
+            {sourcesAndGuardsRenderer(f.source, {
+              messageId: `sg-${f.name}`,
+              streaming: false,
+              role: 'assistant',
+            })}
+          </div>
+
+          <p className="mt-3 rounded-md bg-amber-50 px-3 py-2 text-xs text-amber-900 dark:bg-amber-950/40 dark:text-amber-200">
+            <strong>Guard:</strong> {f.guard}
+          </p>
+        </section>
+      ))}
+    </div>
+  );
+}
+
+export const SourcesAndGuards: Story = {
+  name: 'Sources & Guards',
+  parameters: {
+    docs: {
+      description: {
+        story:
+          'Per-feature documentation of the Markdown **source** that generates each visual and the ' +
+          '**guard** (trust boundary) that sanitizes untrusted model/agent output. Useful for ' +
+          'security review: it shows exactly where each plugin opens the allow-list and how it ' +
+          'degrades.',
+      },
+    },
+  },
+  render: () => <SourcesAndGuardsDemo />,
+};
